Add job.failed&Fix random job.currentDuration on Overview by sorting
When a container can't start because of configuration errors (CreateContainerConfigError),
the container goes into Waiting state. This also needs to be counted and displayed
as a failed cronjob/job/pod.

Since the container didn't fail, is missing some fields as is in waiting mode.
To count it as a failure through the existing code is returned in `failedContainers`
and 'fakes' `TerminatedReason`/`ContainerStateTerminated` are created and returned.